+++ HaveToDO/screens/InformationScreen.tsx
@@ -5,15 +5,13 @@ import { Platform, StyleSheet } from 'react-native';
 import EditScreenInfo from '../components/EditScreenInfo';
 import { Text, View } from '../components/Themed';
 
-export default function ModalScreen() {
+export default function InformationScreen() {
   return (
     <View style={styles.container}>
-      <Text style={styles.title}>Modal</Text>
-      <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
-      <EditScreenInfo path="/screens/ModalScreen.tsx" />
-
-      {/* Use a light status bar on iOS to account for the black space above the modal */}
-      <StatusBar style={Platform.OS === 'ios' ? 'light' : 'auto'} />
+      <Text style={styles.title}>Información</Text>
+      <Text>Somos el Grupo 16,17 y 18 del Ciclo 4A</Text>
+      <Text>Programación web</Text>
+      <Text>MisiónTIC 2022</Text>
     </View>
   );
 }
